Migrating and seeding properly. Adding all files. getAll endpoint functioning

+++ data/migrations/20210909214630_recipe_ingredients_table.js
@@ -10,7 +10,7 @@ exports.up = function (knex) {
       .index()
       .references('ingredient_id')
       .inTable('ingredients')
-    table.integer('quantity')
+    table.float('quantity')
       .notNullable()
   })
 }
